support double precision, spawning setting, PFM, cifar10_w data script

+++ spiking_layer_ours.py
@@ -11,7 +11,9 @@ class SpikeAct(torch.autograd.Function):
     def forward(ctx, mem, spike_index_inner, spike_index_outer, const):
         
         ctx.save_for_backward(mem.clone(), const*spike_index_inner.clone(), spike_index_inner.clone())
-        spike = mem.ge(spike_index_inner.item() * const).float() * spike_index_outer.item() * const
+        spike = mem.ge(spike_index_inner.item() * const).to(spike_index_inner.dtype) * spike_index_outer.item() * const
+        # check if it's TypeError
+            
         return spike
     @staticmethod
     def backward(ctx, grad_output):
@@ -29,19 +31,22 @@ class SPIKE_layer(nn.Module):
         self.thresh_inner = thresh_inner
       
     
-    def forward(self, input,t):
+    def forward(self, input, t):
        
         x = input
         mem = 0
         spike_pot = []
         T = x.shape[1]
+        
+        self.thresh_outer = self.thresh_outer.to(input.dtype)
+        self.thresh_inner = self.thresh_inner.to(input.dtype)
    
         const = 1
         for t in range(T):
             mem += x[:, t, ...]
             spike = SpikeAct.apply(mem, self.thresh_inner[t], self.thresh_outer[t], T)
             # soft-rest
-            mem -= const*mem.ge(self.thresh_inner[t]*T).float() * self.thresh_outer[t] * T
+            mem -= const*mem.ge(self.thresh_inner[t]*T).to(self.thresh_inner[t].dtype) * self.thresh_outer[t] * T
             spike_pot.append(spike)
         return torch.stack(spike_pot, dim=1)
